refactor: move fixture
